+++ R/head.R
@@ -2,7 +2,7 @@
 #'
 #' @description
 #' These functions allow you to view the first (`head()`) or last (`tail()`) `n`
-#' rows of a {gtsummary} table.
+#' rows of a gtsummary table.
 #'
 #' @param x A `gtsummary` object.
 #' @param n Number of rows to return. Default is `6L` as default in [utils::head()] and